Massive asynchronous rewrite using most.js streams.
Where possible, all iterative computations are emitted on streams.
When collated data is necessary for computation to proceed, streams
are reduced into promises, computation is done synchronously, and
next-step data is emitted on a new stream.

Some changes were made beyond adapting the original implementation
to an asynchronous paradigm.  This was particularly true where
mutation was used, or where assumptions were made about the state
of the data that became fase assumptions when made asynchronous.

A couple of minor changes to linting rules, etc, were also added.