Introduces the notion of module providers to the Interlock runtime.
Before this change, there was a single mechanism by which a requested module
could be resolved and loaded onto the page - the `urls` hash was consulted
for a bundle containing the requested module, and a <script> tag was appended
to the DOM. After this change, that is the default/fallback behavior.

Plugins or other JS can now push a function to the
window.__interlock__.providers array.  When a request comes in for a particular
module hash, these providers are invoked in reverse order with two parameters:
the `moduleHash`, and the `next` provider.  It is the responsibility of each
provider to invoke the `next` provider in cases where the provider fails or is
otherwise unable to fulfill the request.

If the `providers` array is exhausted, the invocation chain will terminate
with the throwing of an error, indicating the failure to load module with
the specified hash.

Closes #28.