Replace `worker-farm` with `bson-ipc` multiprocess implementation.
Unfortunately, making these multiprocess changes seems to have made some deeper
bug apparent.  Which all targets enabled, and in certain cases, the
`.then(this.transformModule)` in `src/compile/modules/compile.js#compileModule`
will not fire.

This behavior is not consistent, and appears to be unrelated to either multiprocess,
pluggables, and error in transpilation, or anything else obvious.  It could be
a bug in Bluebird, but it is hard to say for sure.