Nuke `module.type` on watched-file builds.
Without this `loadModule` would check for the module type and return early
and skip reading the file from disk.  This caused watched-file incremental
builds to fail, since `rawSource` would still be null (it _was_ nuked) on
file change and Babel would obviously fail to parse `null` in the `parse`
step.